+++ backend/src/main/java/com/kristianconk/api_papeleria/ventas/VentaPorProductoDTO.java
@@ -1,9 +1,11 @@
 package com.kristianconk.api_papeleria.ventas;
 
+import java.math.BigDecimal;
+
 public record VentaPorProductoDTO(
         Long productoId,
         String productoNombre,
         Integer cantidadVendida,
-        Double montoTotal
+        BigDecimal montoTotal
 ) {
 }
